variational_adaptivity/l_shape_with_singularity/requirements.txt
pandas
variational_adaptivity @ git+ssh://git@github.com/leuraph/variational_adaptivity@v0.0.4
numpy-stl